--- a/src/Jackett.Common/Services/Interfaces/IIndexerConfigurationService.cs
+++ b/src/Jackett.Common/Services/Interfaces/IIndexerConfigurationService.cs
@@ -8,6 +8,6 @@ namespace Jackett.Common.Services.Interfaces
         void Load(IIndexer indexer);
         void Save(IIndexer indexer, JToken config);
         void Delete(IIndexer indexer);
-        string GetIndexerConfigFilePath(string indexerId);
+        void RenameIndexer(string oldId, string newId);
     }
 }