--- a/src/Jackett.Performance/Services/PerformanceProtectionService.cs
+++ b/src/Jackett.Performance/Services/PerformanceProtectionService.cs
@@ -1,0 +1,11 @@
+using Jackett.Common.Services.Interfaces;
+
+namespace Jackett.Performance.Services
+{
+    public class PerformanceProtectionService : IProtectionService
+    {
+        public string Protect(string plainText) => plainText;
+
+        public string UnProtect(string plainText) => plainText;
+    }
+}