--- a/src/Jackett.Performance/Services/PerformanceServerService.cs
+++ b/src/Jackett.Performance/Services/PerformanceServerService.cs
@@ -1,0 +1,14 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Jackett.Common.Models.Config;
+using Jackett.Common.Services.Interfaces;
+using Jackett.Common.Utils.Clients;
+using Jackett.Server.Services;
+using NLog;
+
+namespace Jackett.Performance.Services
+{
+}