@ilike2burnthing 2021-05-13
Apologies for my ignorance, but what is being measured here? The time taken, and memory needed, for an indexer to process a prefetched results page?

This would be useful when making significant changes to the indexer, or Jackett as a whole, to see if there's any significant impact, either positive or negative?