@alecsg77 2021-05-14
> Apologies for my ignorance, but what is being measured here? The time taken, and memory needed, for an indexer to process a prefetched results page?
Yes, correct, or better, this is my goal.
 
> This would be useful when making significant changes to the indexer, or Jackett as a whole, to see if there's any significant impact, either positive or negative?
Yes.