@ilike2burnthing 2024-01-18
Never really go off the ground. It was the proposal of a contributor. Looking at the activity of the user, they took almost a year break from GH after submitting this.